Implement blocking using the filesystem namespace (#30)
* Implement blocking using the filesystem namespace

This implements blocking of IP addresses using the file system namespace.

Resolves #29

* Don't for an implicit `rule block`

I decided my earlier choice to make using `prefix_dir` imply `rule
block` was not a good idea. It is both inconsistent with how the other
subdirectives work and makes it impossible to use this mechanism to
dynamically whitelist addresses. I can't think of a situation where
someone would want to use it that way but that's not a good reason to
make it impossible.

* Add prefix_dir test cases