Require exactly one `rule` directive per block
If a `rule` directive isn't provided it defaults to `allow` which may
confuse people. So require the user to explicitly specify whether the
filter should `block` or `allow`. Similarly, users might be tempted to
specify the `rule` more than once per block which means they have
misunderstood how the ipfilter DSL works. So treat that as an error.

Resolves #31